Move domain to data model mapping out of the aggregate roots and rename model to datamodel in domain to differentiate from database pure models
