+++ app/inventory/internal/vehicle/domain/datamodel/car.go
@@ -1,11 +1,11 @@
-package model
+package datamodel
 
 import (
-	"vehicle-sharing-go/pkg/domain/model"
+	"vehicle-sharing-go/pkg/domain/datamodel"
 )
 
 type Car struct {
 	VinNumber string `gorm:"type:varchar(255);unique"`
 	Color     string `gorm:"type:varchar(255)"`
-	*model.AggregateRoot
+	*datamodel.AggregateRoot
 }
