+++ pkg/domain/datamodel/aggregate_root.go
@@ -1,4 +1,4 @@
-package model
+package datamodel
 
 import (
 	"time"
